// apps/api/src/lib/source/imdbapidev/generated/models/imdbapi-name.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IMDbAPI
 * IMDb API for accessing movie and TV show data
 *
 * The version of the OpenAPI document: 2.7.12
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ImdbapiImage } from './imdbapi-image.js';
// May contain unused imports in some cases
// @ts-ignore
import type { ImdbapiNameMeterRanking } from './imdbapi-name-meter-ranking.js';
// May contain unused imports in some cases
// @ts-ignore
import type { ImdbapiPrecisionDate } from './imdbapi-precision-date.js';

/**
 * The Name message represents a person in the IMDb database, such as an actor, director, or producer.
 * @export
 * @interface ImdbapiName
 */
export interface ImdbapiName {
    /**
     * The unique identifier for the name in the IMDb database.
     * @type {string}
     * @memberof ImdbapiName
     */
    'id'?: string;
    /**
     * The display name of the person, typically their full name.
     * @type {string}
     * @memberof ImdbapiName
     */
    'displayName'?: string;
    /**
     * Alternative names for the person, which may include stage names, nicknames, or other variations.
     * @type {Array<string>}
     * @memberof ImdbapiName
     */
    'alternativeNames'?: Array<string>;
    /**
     * 
     * @type {ImdbapiImage}
     * @memberof ImdbapiName
     */
    'primaryImage'?: ImdbapiImage;
    /**
     * A list of primary professions associated with the person, such as \"Actor\", \"Director\", \"Producer\", etc.
     * @type {Array<string>}
     * @memberof ImdbapiName
     */
    'primaryProfessions'?: Array<string>;
    /**
     * A brief biography or description of the person, which may include their career highlights, achievements, and other relevant information.
     * @type {string}
     * @memberof ImdbapiName
     */
    'biography'?: string;
    /**
     * The height of the person in centimeters.
     * @type {number}
     * @memberof ImdbapiName
     */
    'heightCm'?: number;
    /**
     * The birth name of the person, which may differ from their display name.
     * @type {string}
     * @memberof ImdbapiName
     */
    'birthName'?: string;
    /**
     * 
     * @type {ImdbapiPrecisionDate}
     * @memberof ImdbapiName
     */
    'birthDate'?: ImdbapiPrecisionDate;
    /**
     * The birth location of the person, which may include the city and country of birth.
     * @type {string}
     * @memberof ImdbapiName
     */
    'birthLocation'?: string;
    /**
     * 
     * @type {ImdbapiPrecisionDate}
     * @memberof ImdbapiName
     */
    'deathDate'?: ImdbapiPrecisionDate;
    /**
     * The death location of the person, which may include the city and country of death.
     * @type {string}
     * @memberof ImdbapiName
     */
    'deathLocation'?: string;
    /**
     * The reason for the person\'s death, if applicable.
     * @type {string}
     * @memberof ImdbapiName
     */
    'deathReason'?: string;
    /**
     * 
     * @type {ImdbapiNameMeterRanking}
     * @memberof ImdbapiName
     */
    'meterRanking'?: ImdbapiNameMeterRanking;
}