// apps/api/src/lib/source/imdbapidev/generated/models/imdbapi-title.ts
export interface ImdbapiTitle {
    /**
     * The unique identifier for the title.
     * @type {string}
     * @memberof ImdbapiTitle
     */
    'id'?: string;
    /**
     * The type of the title, such as \"movie\", \"tvSeries\", \"tvEpisode\", etc.
     * @type {string}
     * @memberof ImdbapiTitle
     */
    'type'?: string;
    /**
     * The is_adult field indicates whether the title is intended for adult audiences.
     * @type {boolean}
     * @memberof ImdbapiTitle
     */
    'isAdult'?: boolean;
    /**
     * The primary title of the title, which is typically the most recognized name.
     * @type {string}
     * @memberof ImdbapiTitle
     */
    'primaryTitle'?: string;
    /**
     * The original title of the title, normally which is the title as it was originally released.
     * @type {string}
     * @memberof ImdbapiTitle
     */
    'originalTitle'?: string;
    /**
     * 
     * @type {ImdbapiImage}
     * @memberof ImdbapiTitle
     */
    'primaryImage'?: ImdbapiImage;
    /**
     * The start_year field is used for titles that have a defined start, such as movies or TV series.
     * @type {number}
     * @memberof ImdbapiTitle
     */
    'startYear'?: number;
    /**
     * The end_year field is used for titles that have a defined end, such as TV series.
     * @type {number}
     * @memberof ImdbapiTitle
     */
    'endYear'?: number;
    /**
     * The runtime_seconds field contains the total runtime of the title in minutes.
     * @type {number}
     * @memberof ImdbapiTitle
     */
    'runtimeSeconds'?: number;
    /**
     * The genres field contains a list of genres associated with the title.
     * @type {Array<string>}
     * @memberof ImdbapiTitle
     */
    'genres'?: Array<string>;
    /**
     * 
     * @type {ImdbapiRating}
     * @memberof ImdbapiTitle
     */
    'rating'?: ImdbapiRating;
    /**
     * 
     * @type {ImdbapiMetacritic}
     * @memberof ImdbapiTitle
     */
    'metacritic'?: ImdbapiMetacritic;
    /**
     * The plot field contains a brief summary or description of the title\'s storyline.
     * @type {string}
     * @memberof ImdbapiTitle
     */
    'plot'?: string;
    /**
     * The list of directors associated with the title, which can include multiple directors.
     * @type {Array<ImdbapiName>}
     * @memberof ImdbapiTitle
     */
    'directors'?: Array<ImdbapiName>;
    /**
     * The list of writers associated with the title, which can include multiple writers.
     * @type {Array<ImdbapiName>}
     * @memberof ImdbapiTitle
     */
    'writers'?: Array<ImdbapiName>;
    /**
     * The list of stars associated with the title, which can include multiple actors or actresses.
     * @type {Array<ImdbapiName>}
     * @memberof ImdbapiTitle
     */
    'stars'?: Array<ImdbapiName>;
    /**
     * The list of countries where the title was produced or is associated with.
     * @type {Array<ImdbapiCountry>}
     * @memberof ImdbapiTitle
     */
    'originCountries'?: Array<ImdbapiCountry>;
    /**
     * The list of languages spoken in the title, which can include multiple languages.
     * @type {Array<ImdbapiLanguage>}
     * @memberof ImdbapiTitle
     */
    'spokenLanguages'?: Array<ImdbapiLanguage>;
    /**
     * The list of interests associated with the title.
     * @type {Array<ImdbapiInterest>}
     * @memberof ImdbapiTitle
     */
    'interests'?: Array<ImdbapiInterest>;
}